Backend: Fix restaurant_results and Alembic migration
