--- alembic/versions/594f00541569_add_restaurant_results_table.py
@@ -1,30 +1,0 @@
-"""Add restaurant_results table
-
-Revision ID: 594f00541569
-Revises: c4eb24204e2c
-Create Date: 2025-07-26 06:48:26.519293
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '594f00541569'
-down_revision: Union[str, None] = 'c4eb24204e2c'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
