+++ alembic/versions/9e0f97e315d0_add_restaurant_results_table.py
@@ -1,0 +1,42 @@
+"""Add restaurant_results table
+
+Revision ID: 9e0f97e315d0
+Revises: c4eb24204e2c
+Create Date: 2025-07-27 01:17:54.491689
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '9e0f97e315d0'
+down_revision: Union[str, None] = 'c4eb24204e2c'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('restaurant_results',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('user_id', sa.Integer(), nullable=True),
+    sa.Column('restaurant_name', sa.String(), nullable=True),
+    sa.Column('meal_name', sa.String(), nullable=True),
+    sa.Column('macros', sa.JSON(), nullable=True),
+    sa.Column('explanation', sa.String(), nullable=True),
+    sa.Column('timestamp', sa.String(), nullable=True),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_restaurant_results_id'), 'restaurant_results', ['id'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_restaurant_results_id'), table_name='restaurant_results')
+    op.drop_table('restaurant_results')
+    # ### end Alembic commands ###
