+++ backend/app/models/restaurant.py
@@ -6,7 +6,7 @@ class RestaurantResult(Base):
     __tablename__ = "restaurant_results"
 
     id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, ForeignKey("user.id"))
+    user_id = Column(Integer, ForeignKey("users.id"))
     restaurant_name = Column(String)
     meal_name = Column(String)
     macros = Column(JSON)  # example: {"calories": 500, "protein": 40, "carbs": 50, "fats": 18, "fiber": 8}
